requirements.txt
cachetools
click==8.0.4
Flask==2.2.3
google-auth # ==2.5.0
google-cloud-datastore==1.15.3
google-cloud-ndb==1.11.1
google-cloud-tasks==2.13.0
requests==2.27.1
charset_normalizer==2.0.11
docx-mailmerge==0.5.0
openpyxl==3.1.1
jsonpickle==3.0.1
# https://github.com/GoogleCloudPlatform/appengine-python-standard
appengine-python-standard>=1.0.0